+++ projects/ad74413r/Makefile
@@ -1,9 +1,7 @@
-# Select the example you want to enable by choosing y for enabling and n for disabling
-DUMMY_EXAMPLE = y
-IIO_EXAMPLE = n
-IIO_TRIGGER_EXAMPLE = n
-
 include ../../tools/scripts/generic_variables.mk
+
+EXAMPLE := dummy
+include ../../tools/scripts/examples.mk
 
 include src.mk
