+++ projects/ad74413r/src/examples/dummy/dummy_example.c
@@ -34,7 +34,6 @@
 /******************************************************************************/
 /***************************** Include Files **********************************/
 /******************************************************************************/
-#include "dummy_example.h"
 #include "common_data.h"
 #include "ad74413r.h"
 #include "no_os_delay.h"
@@ -49,7 +48,7 @@
  * @return ret - Result of the example execution. If working correctly, will
  *               execute continuously the while(1) loop and will not return.
 *******************************************************************************/
-int dummy_example_main()
+int example_main()
 {
 	int ret;
 	uint8_t val;
@@ -63,6 +62,14 @@ int dummy_example_main()
 	};
 
 	struct ad74413r_decimal result[3];
+
+	struct no_os_uart_desc *uart_desc;
+
+	ret = no_os_uart_init(&uart_desc, &ad74413r_uart_ip);
+	if (ret)
+		return ret;
+
+	no_os_uart_stdio(uart_desc);
 
 	ret = ad74413r_init(&ad74413r_desc, &ad74413r_ip);
 	if (ret)
