--- projects/ad74413r/src/examples/examples_src.mk
@@ -1,32 +1,0 @@
-ifeq (y,$(strip $(DUMMY_EXAMPLE)))
-CFLAGS += -DDUMMY_EXAMPLE
-SRCS += $(PROJECT)/src/examples/dummy/dummy_example.c
-INCS += $(PROJECT)/src/examples/dummy/dummy_example.h
-endif
-
-ifeq (y,$(strip $(IIO_EXAMPLE)))
-IIOD=y
-CFLAGS += -DIIO_EXAMPLE
-SRCS += $(PROJECT)/src/examples/iio_example/iio_example.c
-INCS += $(PROJECT)/src/examples/iio_example/iio_example.h
-endif
-
-ifeq (y,$(strip $(IIO_TRIGGER_EXAMPLE)))
-IIOD=y
-CFLAGS += -DIIO_TRIGGER_EXAMPLE
-SRCS += $(PROJECT)/src/examples/iio_trigger_example/iio_trigger_example.c
-INCS += $(PROJECT)/src/examples/iio_trigger_example/iio_trigger_example.h
-endif
-
-ifeq (y,$(strip $(IIOD)))
-SRC_DIRS += $(NO-OS)/iio/iio_app
-
-INCS += $(DRIVERS)/adc-dac/ad74413r/iio_ad74413r.h
-SRCS += $(DRIVERS)/adc-dac/ad74413r/iio_ad74413r.c
-
-SRCS += $(NO-OS)/iio/iio_trigger.c
-INCS += $(NO-OS)/iio/iio_trigger.h
-
-INCS += $(INCLUDE)/no_os_list.h \
-	$(PLATFORM_DRIVERS)/$(PLATFORM)_uart.h
-endif
