+++ projects/ad74413r/src/examples/iio_trigger_example/example.mk
@@ -1,0 +1,4 @@
+IIOD=y
+
+INCS += $(DRIVERS)/adc-dac/ad74413r/iio_ad74413r.h
+SRCS += $(DRIVERS)/adc-dac/ad74413r/iio_ad74413r.c
