+++ projects/ad74413r/src/platform/stm32/platform.mk
